+++ CHANGELOG.md
@@ -4,6 +4,7 @@ The version is of the bottom crate `truck-rendimpl`.
 
 ## Unreleased
 
+- Implement `Offset` geometry.
 - Remove render object by `RenderID`.
 - Fix example pages for working on Mac.
 - Make all `proptest!` `#[property_test]`.